+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 - Unify objects task [#970]
 - Attribution to ObservationMatrix
+- New biological association task [#4026]
 
 ### Changed
 
@@ -21,10 +22,11 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Containerizing objects prevented identific increments
 - New dichotomous key: Radial annotator is not loading the correct data [#4076]
 
-[#4077]: https://github.com/SpeciesFileGroup/taxonworks/issues/4077
-[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
+[#4026]: https://github.com/SpeciesFileGroup/taxonworks/issues/4026
 [#4074]: https://github.com/SpeciesFileGroup/taxonworks/issues/4074
 [#4076]: https://github.com/SpeciesFileGroup/taxonworks/issues/4076
+[#4077]: https://github.com/SpeciesFileGroup/taxonworks/issues/4077
+[#970]: https://github.com/SpeciesFileGroup/taxonworks/issues/970
 
 ## [0.44.3] - 2024-10-03